requirements.txt
# Python邮件群发工具依赖
pandas>=1.3.0
openpyxl>=3.0.0
python-dotenv>=0.19.0

# 可选依赖（用于增强功能）
# beautifulsoup4>=4.9.3  # 用于处理HTML内容
# jinja2>=3.0.0          # 用于模板渲染